test(core): migrate 8 context tests jvmTest→commonTest, fix inject import
Migrates ContextTest, ContextTypeParameterizationTest, TransformationValidationTest,
ContextTransformationPhasesTest, ContextTransformationTest, BaseContextTest,
ContextImmutabilityTest, ContextInheritanceTest from jvmTest to commonTest.

Fixes: wrong `org.koin.test.inject` → `org.koin.core.component.inject` in all
context commonTest files (KoinComponent, not KoinTest), grid size 30→100 in
CommonTestEditingContextFactory, restores TestContextBuilder factory binding,
adds `override` to SimulationContextFactory methods, and uses afterKoinSetUp()
in ContextTest instead of @BeforeTest to avoid native ordering undefined behaviour.

jvmTest: 1864 tests, 0 failed | linuxX64: 746 tests, 0 failed
