+++ core/src/commonTest/kotlin/cz/vutbr/fit/interlockSim/context/BaseContextTest.kt
@@ -19,7 +19,7 @@ import cz.vutbr.fit.interlockSim.objects.tracks.SimpleTrackBlock
 import cz.vutbr.fit.interlockSim.testutil.CommonKoinTestBase
 import cz.vutbr.fit.interlockSim.util.Point
 import kotlin.test.Test
-import org.koin.test.inject
+import org.koin.core.component.inject
 import cz.vutbr.fit.interlockSim.objects.core.ContextPropertyChangeListener
 
 /**
