+++ core/src/commonTest/kotlin/cz/vutbr/fit/interlockSim/context/ContextInheritanceTest.kt
@@ -16,7 +16,7 @@ import assertk.assertions.isNotInstanceOf
 import assertk.assertions.isTrue
 import cz.vutbr.fit.interlockSim.testutil.CommonKoinTestBase
 import kotlin.test.Test
-import org.koin.test.inject
+import org.koin.core.component.inject
 
 /**
  * Test suite for context inheritance structure.
