+++ core/src/commonTest/kotlin/cz/vutbr/fit/interlockSim/context/ContextSerializationTest.kt
@@ -21,7 +21,7 @@ import cz.vutbr.fit.interlockSim.objects.core.ContextPropertyChangeListener
 import cz.vutbr.fit.interlockSim.testutil.CommonKoinTestBase
 import cz.vutbr.fit.interlockSim.util.Point
 import kotlin.test.Test
-import org.koin.test.inject
+import org.koin.core.component.inject
 
 class ContextSerializationTest : CommonKoinTestBase() {
 	private val editingContextFactory: EditingContextFactory by inject()
