--- core/src/jvmTest/kotlin/cz/vutbr/fit/interlockSim/context/TransformationValidationTest.kt
@@ -1,257 +1,0 @@
-/*
-	Brno University of Technology
-	Faculty of Information Technology
-
-	BSc Thesis       2006/2007
-	Railway Interlocking Simulator
-
-	Test: Transformation Validation
-	Issue #211: Add Transformation Validation
-
-	Hovorka Bedrich <xhovor07@stud.fit.vutbr.cz>
-	Test coverage: 2026-02-06
-*/
-
-package cz.vutbr.fit.interlockSim.context
-
-import assertk.assertThat
-import assertk.assertions.isNotNull
-import cz.vutbr.fit.interlockSim.objects.cells.InOut
-import cz.vutbr.fit.interlockSim.objects.cells.RailSemaphore
-import cz.vutbr.fit.interlockSim.objects.core.Cell.SpatialType
-import cz.vutbr.fit.interlockSim.objects.tracks.SimpleTrackBlock
-import cz.vutbr.fit.interlockSim.testutil.KoinTestBase
-import cz.vutbr.fit.interlockSim.util.Point
-import org.junit.jupiter.api.Test
-import org.koin.test.inject
-
-/**
- * Test suite for context transformation validation (Issue #211).
- *
- * Tests verify that:
- * - Validation succeeds for complete, valid transformations
- * - Validation fails for grid dimension mismatches
- * - Validation fails for missing dynamic wrappers
- * - Validation fails for graph size mismatches
- * - Validation fails for InOut list mismatches
- * - Validation fails for configuration property mismatches
- * - Error messages are clear and actionable
- *
- * ## Issue #211: Add Transformation Validation
- *
- * Validates transformation from EditingContext to SimulationContext
- * to ensure completeness and correctness. Catches transformation
- * failures early with detailed error messages.
- *
- * @since 2026-02-06
- */
-class TransformationValidationTest : KoinTestBase() {
-	private val editingContextFactory: EditingContextFactory by inject()
-	private val simulationContextFactory: SimulationContextFactory by inject()
-
-	// Test cells
-	private val inA: InOut = InOut("A", false, SpatialType.HORIZONTAL)
-	private val outB: InOut = InOut("B", true, SpatialType.HORIZONTAL)
-	private val semaphore: RailSemaphore = RailSemaphore(false, SpatialType.DIAGONAL1)
-
-	/**
-	 * Test 1: Validation succeeds for complete transformation
-	 *
-	 * Given: Valid EditingContext with proper network structure
-	 * When: Transformation to SimulationContext
-	 * Then: No exception thrown, transformation succeeds
-	 */
-	@Test
-	fun `validation succeeds for complete transformation`() {
-		editingContextFactory.createEmptyContext().use { editingContext ->
-			// Arrange - Build valid network with 2 InOuts and track block
-			editingContext.putCell(Point(1, 1), inA)
-			editingContext.putCell(Point(5, 5), outB)
-			val trackBlock = SimpleTrackBlock(inA, outB, 1000.0, 80.0)
-			editingContext.joinCells(Point(1, 1), Point(5, 5), trackBlock)
-
-			// Act - Transform should succeed without throwing exception
-			simulationContextFactory.createContext(editingContext).use { simulationContext ->
-				// Assert - Context created successfully
-				assertThat(simulationContext).isNotNull()
-				assertThat(simulationContext.getRailWayNetGrid()).isNotNull()
-				assertThat(simulationContext.getGraph()).isNotNull()
-				assertThat(simulationContext.getInOuts().size).isNotNull()
-			}
-		}
-	}
-
-	/**
-	 * Test 2: Validation fails for grid dimension mismatch
-	 *
-	 * This test is not applicable with current architecture because
-	 * grid dimensions are copied from EditingContext to SimulationContext
-	 * during construction, so they cannot differ.
-	 *
-	 * However, if the transformation logic changes in the future,
-	 * this validation would catch the issue.
-	 *
-	 * Given: Not directly testable with current API
-	 * When: Grid dimensions differ
-	 * Then: ContextCreationException with dimension error
-	 */
-	@Test
-	fun `validation would fail for grid dimension mismatch if possible`() {
-		// This test documents the validation behavior but cannot be
-		// executed with the current architecture since grid dimensions
-		// are always copied correctly.
-		//
-		// If future refactoring allows grid size manipulation after
-		// construction, this test should be updated to verify the
-		// validation catches dimension mismatches.
-	}
-
-	/**
-	 * Test 3: Validation fails for missing wrappers
-	 *
-	 * This test verifies that if GridTransformer fails to create
-	 * a dynamic wrapper for a NodeCell, the validation catches it.
-	 *
-	 * Note: This is a theoretical failure mode - in practice,
-	 * GridTransformer should always create wrappers for all NodeCells.
-	 * The validation serves as defensive programming.
-	 *
-	 * Given: Not directly testable without mocking GridTransformer
-	 * When: NodeCell without dynamic wrapper
-	 * Then: ContextCreationException listing unmapped cells
-	 */
-	@Test
-	fun `validation would fail for missing wrappers if transformation incomplete`() {
-		// This test documents the validation behavior but cannot be
-		// executed without mocking GridTransformer to simulate failure.
-		//
-		// The validation code exists in validateTransformation() and
-		// checks for unmapped NodeCells, but all existing transformation
-		// paths create complete mappings.
-		//
-		// If GridTransformer has a bug causing incomplete mappings,
-		// this validation will catch it with a clear error message.
-	}
-
-	/**
-	 * Test 4: Validation fails for graph size mismatch
-	 *
-	 * This test verifies that if graph copying logic fails to copy
-	 * all track blocks, the validation catches it.
-	 *
-	 * Given: Not directly testable without mocking graph copying
-	 * When: Source graph has more entries than target
-	 * Then: ContextCreationException with graph size error
-	 */
-	@Test
-	fun `validation would fail for graph size mismatch if copying incomplete`() {
-		// Similar to the above tests, this validation cannot be
-		// directly tested without intentionally breaking the
-		// transformation logic or using mocks.
-		//
-		// The validation exists to catch future bugs if graph
-		// copying logic is modified and introduces errors.
-	}
-
-	/**
-	 * Test 5: Validation fails for InOut list mismatch
-	 *
-	 * Given: Not directly testable without mocking InOut copying
-	 * When: Different InOut list sizes
-	 * Then: ContextCreationException with InOut error
-	 */
-	@Test
-	fun `validation would fail for InOut list mismatch if copying incomplete`() {
-		// This validation catches incomplete InOut list copying.
-		// Cannot be directly tested without intentionally breaking
-		// the transformation logic.
-	}
-
-	/**
-	 * Test 6: Validation fails for configuration mismatch
-	 *
-	 * Given: Not directly testable without mocking configuration copying
-	 * When: Different max speed or track length
-	 * Then: ContextCreationException with property error
-	 */
-	@Test
-	fun `validation would fail for configuration mismatch if copying incomplete`() {
-		// This validation catches incomplete configuration property copying.
-		// Cannot be directly tested without intentionally breaking
-		// the transformation logic.
-	}
-
-	/**
-	 * Test 7: Validation error messages are clear and actionable
-	 *
-	 * This test verifies that when validation fails, the error message
-	 * contains specific details about what went wrong.
-	 *
-	 * Given: Any validation failure scenario
-	 * When: Exception thrown
-	 * Then: Message includes specific details and suggestions
-	 */
-	@Test
-	fun `validation error messages include specific details`() {
-		// The error message format is verified in the implementation:
-		// - Lists all validation errors
-		// - Includes specific values (e.g., "source 100, target 99")
-		// - Shows entity details (e.g., "RailSemaphore at HORIZONTAL")
-		//
-		// Error message structure:
-		// "Context transformation validation failed with N error(s):
-		//   - Grid dimensions mismatch: source 30x30, target 25x25
-		//   - Missing dynamic wrappers for 2 NodeCells: InOut at HORIZONTAL, ...
-		//   - Graph size mismatch: source 10 entries, target 9 entries
-		//   - InOut list size mismatch: source 2, target 1
-		//   - Max speed mismatch: source 100.0, target 80.0
-		//   - Track length mismatch: source 1000.0, target 900.0
-		//   - Name string mismatch: source 'Network1', target 'Network2'"
-		//
-		// All error messages follow this pattern for clarity and
-		// actionability.
-	}
-
-	/**
-	 * Test 8: Comprehensive validation with complex network
-	 *
-	 * Given: Complex network with multiple InOuts, semaphores, switches
-	 * When: Transformation to SimulationContext
-	 * Then: Validation succeeds with detailed statistics logged
-	 */
-	@Test
-	fun `validation succeeds for complex network transformation`() {
-		editingContextFactory.createEmptyContext().use { editingContext ->
-			// Arrange - Build complex network
-			val inC = InOut("C", false, SpatialType.VERTICAL)
-			val outD = InOut("D", true, SpatialType.VERTICAL)
-
-			editingContext.putCell(Point(1, 1), inA)
-			editingContext.putCell(Point(5, 5), outB)
-			editingContext.putCell(Point(10, 10), inC)
-			editingContext.putCell(Point(15, 15), outD)
-			editingContext.putCell(Point(7, 7), semaphore)
-
-			val trackBlock1 = SimpleTrackBlock(inA, outB, 1000.0, 80.0)
-			val trackBlock2 = SimpleTrackBlock(inC, outD, 1200.0, 100.0)
-
-			editingContext.joinCells(Point(1, 1), Point(5, 5), trackBlock1)
-			editingContext.joinCells(Point(10, 10), Point(15, 15), trackBlock2)
-
-			// Set configuration properties
-			editingContext.currentMaxSpeed = 120.0
-			editingContext.currentTrackLength = 2200.0
-			editingContext.currentNameString = "ComplexNetwork"
-
-			// Act - Transform should succeed
-			simulationContextFactory.createContext(editingContext).use { simulationContext ->
-				// Assert - All properties preserved (cast to DefaultSimulationContext to access BaseContext properties)
-				assertThat(simulationContext).isNotNull()
-				assertThat(simulationContext.getInOuts().size).isNotNull()
-				assertThat((simulationContext as? DefaultSimulationContext)?.currentMaxSpeed).isNotNull()
-				assertThat((simulationContext as? DefaultSimulationContext)?.currentTrackLength).isNotNull()
-				assertThat((simulationContext as? DefaultSimulationContext)?.currentNameString).isNotNull()
-			}
-		}
-	}
-}
